Move AggregateType to the top level
