+++ h2/src/main/org/h2/expression/aggregate/Aggregate.java
@@ -48,113 +48,6 @@ import org.h2.value.ValueString;
  */
 public class Aggregate extends AbstractAggregate {
 
-    public enum AggregateType {
-        /**
-         * The aggregate type for COUNT(*).
-         */
-        COUNT_ALL,
-
-        /**
-         * The aggregate type for COUNT(expression).
-         */
-        COUNT,
-
-        /**
-         * The aggregate type for GROUP_CONCAT(...).
-         */
-        GROUP_CONCAT,
-
-        /**
-         * The aggregate type for SUM(expression).
-         */
-        SUM,
-
-        /**
-         * The aggregate type for MIN(expression).
-         */
-        MIN,
-
-        /**
-         * The aggregate type for MAX(expression).
-         */
-        MAX,
-
-        /**
-         * The aggregate type for AVG(expression).
-         */
-        AVG,
-
-        /**
-         * The aggregate type for STDDEV_POP(expression).
-         */
-        STDDEV_POP,
-
-        /**
-         * The aggregate type for STDDEV_SAMP(expression).
-         */
-        STDDEV_SAMP,
-
-        /**
-         * The aggregate type for VAR_POP(expression).
-         */
-        VAR_POP,
-
-        /**
-         * The aggregate type for VAR_SAMP(expression).
-         */
-        VAR_SAMP,
-
-        /**
-         * The aggregate type for ANY(expression).
-         */
-        ANY,
-
-        /**
-         * The aggregate type for EVERY(expression).
-         */
-        EVERY,
-
-        /**
-         * The aggregate type for BOOL_OR(expression).
-         */
-        BIT_OR,
-
-        /**
-         * The aggregate type for BOOL_AND(expression).
-         */
-        BIT_AND,
-
-        /**
-         * The aggregate type for SELECTIVITY(expression).
-         */
-        SELECTIVITY,
-
-        /**
-         * The aggregate type for HISTOGRAM(expression).
-         */
-        HISTOGRAM,
-
-        /**
-         * The aggregate type for MEDIAN(expression).
-         */
-        MEDIAN,
-
-        /**
-         * The aggregate type for ARRAY_AGG(expression).
-         */
-        ARRAY_AGG,
-
-        /**
-         * The aggregate type for MODE(expression).
-         */
-        MODE,
-
-        /**
-         * The aggregate type for ENVELOPE(expression).
-         */
-        ENVELOPE,
-    }
-
     private static final HashMap<String, AggregateType> AGGREGATES = new HashMap<>(64);
 
     private final AggregateType aggregateType;
