+++ h2/src/main/org/h2/expression/aggregate/AggregateData.java
@@ -7,7 +7,6 @@ package org.h2.expression.aggregate;
 
 import org.h2.engine.Constants;
 import org.h2.engine.Database;
-import org.h2.expression.aggregate.Aggregate.AggregateType;
 import org.h2.message.DbException;
 import org.h2.value.Value;
